% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auto}
\alias{auto}
\title{Automotive Industry Brand Perception}
\format{A data frame with 39 rows and 63 variables:
\describe{
  \item{Brand_Name}{Brand Name}
  \item{Category_Name}{Category (automotive)}
  }}
\usage{
auto
}
\description{
A dataset containing brand names, industry type and perception levels on different measures.
}
\keyword{datasets}
