% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny_airbnb}
\alias{ny_airbnb}
\title{NYC Airbnb Data}
\format{A data frame with 38306 rows and 23 variables:
\describe{
  \item{city}{City of Listing}
  \item{state}{State of Listing}
  }}
\usage{
ny_airbnb
}
\description{
NYC Airbnb Data
}
\keyword{datasets}
