% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soup}
\alias{soup}
\title{soup}
\format{A data frame with 88409 rows and 34 variables:
\describe{
  \item{IRI_KEY}{Unique ID}
  \item{Winter}{Variable defining "Winter" or "nonwinter" status}
  }}
\usage{
soup
}
\description{
Store sales data of Progresso and Campbell's soup, combined with demographic information.
}
\keyword{datasets}
