% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soup_store}
\alias{soup_store}
\title{soup_store}
\format{A data frame with 2055 rows and 22 variables:
\describe{
  \item{IRI_KEY}{Unique ID}
  \item{FIPS}{Location FIPS code}
  }}
\usage{
soup_store
}
\description{
Store level soup sales data
}
\keyword{datasets}
