% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sqft}
\alias{sqft}
\title{sqft}
\format{A data frame with 146333 rows and 6 variables:
\describe{
  \item{RegionID}{Unique region identifier}
  \item{RegionName}{Name of region}
  }}
\source{
\url{https://www.zillow.com/research/data/}
}
\usage{
sqft
}
\description{
Median home prices per square foot
}
\keyword{datasets}
